+++ backend/app/main.py
@@ -3,7 +3,7 @@ this script is the entry point for the FastAPI application.
 """
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from backend.app.routers import courses
+from backend.app.routers import courses, requirements
 
 
 app = FastAPI()
@@ -17,3 +17,4 @@ app.add_middleware(
 )
 
 app.include_router(courses.router)
+app.include_router(requirements.router)
