+++ backend/app/routers/courses.py
@@ -10,8 +10,7 @@ from sqlalchemy.orm import Session
 from backend.database.db import get_db
 from backend.services.courses import CourseService
 from backend.app.schemas import (CourseResponse, CourseListResponse,
-                                 CourseFilter, RequirementsResponse,
-                                 DepartmentListResponse, RequirementResponse)
+                                 CourseFilter, DepartmentListResponse)
 
 router = APIRouter()
 
@@ -67,23 +66,9 @@ def get_course(course_code: str, course_service: CourseService = Depends(get_cou
         raise HTTPException(status_code=404, detail="Course not found")
     return course
 
-@router.get("/requirements", response_model=RequirementsResponse)
-def get_requirements(course_service: CourseService = Depends(get_course_service)):
-    """API route to fetch all course requirements with type and major."""
-    return course_service.fetch_all_requirements()
-
 @router.get("/departments", response_model=DepartmentListResponse)
 def get_departments(course_service: CourseService = Depends(get_course_service)):
     """
     API route to fetch all available departments.
     """
     return DepartmentListResponse(departments=course_service.fetch_all_departments())
-
-@router.get("/requirements/{requirement_name}", response_model=RequirementResponse)
-def get_requirement(requirement_name: str,
-                    course_service: CourseService = Depends(get_course_service)):
-    """API route to fetch a single requirement with its type and major."""
-    requirement = course_service.fetch_requirement(requirement_name)
-    if not requirement:
-        raise HTTPException(status_code=404, detail="Requirement not found")
-    return requirement
